Increased the x y c dialog box sizes which caused functions to fail
The server_3 and server_4 functions appear to be failing upon further
debugging when dropping into stage 3 of the main dialog menu the dialog
boxes appeared to be restricted concerning what text they needed to
populate with the space for the x y z arguments; increasing this to 25 75 25
stopped it from occuring.

Having the variables availble in the functions scripts is helpful for
debugging because you can call single functions to see how they are
exiting. `# bash -x ./rhel-labs-functions server 3`.